+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/ServerPools/ServerPools.cs
@@ -181,12 +181,12 @@ namespace Stormancer.Server.Plugins.GameSession.ServerPool
                 var client = await _eSClientFactory.CreateClient<GameServerRecord>("gameservers");
 
                 var record = client.Get<GameServerRecord>(serverId);
-                if (record != null)
+                if (record.Source != null)
                 {
                     record.Source.ClosedOn = DateTime.UtcNow;
                 }
                 await pool.OnGameServerDisconnected(serverId,record?.Source);
-                if (record != null)
+                if (record.Source != null)
                 {
                     record.Source.RunTimeInSeconds = (record.Source.ClosedOn - record.Source.StartedOn).Seconds;
                     await client.IndexAsync(record.Source, desc => desc.Id(record.Id));