+++ src/Stormancer.Plugins/Regions/Stormancer.Server.Plugins.Regions/RegionTestingStorage.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Stormancer.Server.Plugins.Regions
+{
+    public class RegionTestingStorage
+    {
+        public MemoryCache<int, Dictionary<string, string>> Cache { get; } = new MemoryCache<int, Dictionary<string, string>>();
+    }
+}