+++ src/Stormancer.Plugins/Regions/Stormancer.Server.Plugins.Regions/RegionsConfigurationSection.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Stormancer.Server.Plugins.Regions
+{
+    /// <summary>
+    /// Regions configuration section
+    /// </summary>
+    public class RegionsConfigurationSection
+    {
+        /// <summary>
+        /// Default region
+        /// </summary>
+        public string? DefaultRegion { get; set; }
+    }
+}