+++ src/Stormancer.Plugins/GameSession/Stormancer.GameServers.Agent/Changelog.rst
@@ -15,6 +15,7 @@ Unreleased
 Added
 *****
 - Create server crash reports and store them locally.
+- Added APIs and systems to support best region detection.
 
 0.3.0.8
 -------